requirements.txt
fastapi
uvicorn
textstat
pyspellchecker
spacy
firebase-admin
pandas
joblib
scikit-learn
pycountry
pydantic[email]
pdfminer
PyPDF2
python-multipart
